fix remaining broken tests, alter path to patched wizards
