+++ deet/extractors/cli_helpers.py
@@ -14,7 +14,9 @@ from deet.data_models.project import ExperimentArtefacts
 from deet.extractors.llm_data_extractor import DataExtractionConfig
 from deet.processors.linker import DocumentReferenceLinker, LinkingStrategy
 from deet.ui import fail_with_message, notify
-from deet.ui.terminal.wizards import run_model_wizard
+from deet.ui.terminal import console, render_template
+from deet.ui.terminal.components import info_panel
+from deet.ui.terminal.wizards import continue_after_key, run_model_wizard
 
 
 def load_config_from_context(
@@ -29,6 +31,14 @@ def load_config_from_context(
                 "from a project directory, or provide a config file."
             )
             fail_with_message(no_config)
+        console.clear()
+        console.print(
+            info_panel(
+                render_template("extraction/config_init"),
+                "Data extraction config wizard",
+            )
+        )
+        continue_after_key()
         return run_model_wizard(DataExtractionConfig)
     try:
         return DataExtractionConfig.from_yaml(config_path)
