+++ tests/conftest.py
@@ -1,6 +1,9 @@
+import json
 from pathlib import Path
 
 import pytest
+
+from deet.processors.converter_register import SupportedImportFormat
 
 
 @pytest.fixture
@@ -90,4 +93,19 @@ def sample_eppi_data() -> dict:
                 ],
             }
         ],
+    }
+
+
+@pytest.fixture
+def valid_project_data(tmp_path, sample_eppi_data):
+    # Create a real dummy file so Pydantic's FilePath is happy
+    data_file = tmp_path / "data.json"
+    data_file.write_text(json.dumps(sample_eppi_data))
+
+    return {
+        "name": "TestProject",
+        "gold_standard_data_path": data_file,
+        "gold_standard_data_format": SupportedImportFormat.EPPI_JSON,
+        "environment_file": "project",
+        "pdf_dir": tmp_path,  # A real directory
     }
